fix(controller): clean lint debt in recovery_policy.go
Two issues from upstream commit ea3a1da (US-24.5 Recovery policy)
were blocking pre-commit lint:

1) gosec G115: int -> int32 conversion at line 46. The
   SafeModeAfter values are tiny integers (3, 6) so the conversion
   is harmless, but the cleanest fix is to declare the field as
   int32 directly so no conversion is needed. Updates 3 test
   assertions to match.

2) unused: enterRecovery method has no caller yet — it is wired
   up by the failure-class dispatch in US-24.6 (next commit in the
   epic). Add nolint:unused with that pointer rather than deleting
   the WIP method.

Tests still pass (8 recovery_policy_test cases, including the
three updated to assert int32).

+++ controller/internal/workspace/recovery_policy.go
@@ -17,7 +17,7 @@ type RecoveryPolicy struct {
 	BackoffMax     time.Duration
 	BackoffFactor  int
 	StabilityReset time.Duration
-	SafeModeAfter  int
+	SafeModeAfter  int32
 }
 
 var recoveryPolicies = map[FailureClass]RecoveryPolicy{
@@ -43,9 +43,14 @@ func calculateBackoff(failures int32, policy RecoveryPolicy) time.Duration {
 }
 
 func shouldEnterSafeMode(consecutiveFailures int32, policy RecoveryPolicy) bool {
-	return policy.SafeModeAfter > 0 && consecutiveFailures >= int32(policy.SafeModeAfter)
+	return policy.SafeModeAfter > 0 && consecutiveFailures >= policy.SafeModeAfter
 }
 
+// enterRecovery is wired up by the failure-class dispatch in US-24.6.
+// Suppressing the unused warning until the call site lands; the function
+// is exercised by recovery_policy_test.go in the meantime.
+//
+//nolint:unused // wired up by US-24.6 follow-up commit
 func (r *WorkspaceReconciler) enterRecovery(ctx context.Context, ws *v1.Workspace, class FailureClass) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
